feat: switch from unsigned long to uint32_t
The previous use of unsigned longs might be less portable than using inttypes.h and uint32_t. This eliminates the former in favor of the latter.

+++ src/basic-io.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <inttypes.h>
 
 #define PROMPT_ENUS "Please provide some input...\n"
 #define CHAR_COUNT_TEXT_ENUS "You entered the following number of characters"
@@ -8,11 +9,11 @@ int main() {
     /* define a variable in which to store user-provided characters */
     int character;
     /* define a variable in which to store the character count as it grows */
-    unsigned long character_count;
+    uint32_t character_count;
     /* set the initial character count to zero */
     character_count = 0;
     /* define a variable in which to store the line count as it grows */
-    unsigned long line_count;
+    uint32_t line_count;
     /* set the initial line count to zero */
     line_count = 0;
 
@@ -31,8 +32,8 @@ int main() {
     }
 
     /* tell the user how many characters they input */
-    printf("%s: %lu\n", CHAR_COUNT_TEXT_ENUS, character_count);
+    printf("%s: %"PRIu32"\n", CHAR_COUNT_TEXT_ENUS, character_count);
 
     /* tell the user how many lines they input */
-    printf("%s: %lu\n", LINE_COUNT_TEXT_ENUS, line_count);
+    printf("%s: %"PRIu32"\n", LINE_COUNT_TEXT_ENUS, line_count);
 }
